Refactor: move common types to a separate header.
